Fix map struct

+++ src/main/java/com/project1/user/User.java
@@ -26,8 +26,10 @@ public class User implements UserDetails {
 
   @Id
   @GeneratedValue
-  private Long id;
+  private Integer id;
   private String firstname;
+
+
   private String lastname;
 
   @Column(unique = true)
